+++ src/main/java/net/funkpla/nether_tweaks/NetherTweaksConfig.java
@@ -1,4 +1,4 @@
-package net.funkpla.fortress_tweak;
+package net.funkpla.nether_tweaks;
 
 import me.shedaniel.autoconfig.ConfigData;
 import me.shedaniel.autoconfig.annotation.Config;
@@ -7,8 +7,8 @@ import me.shedaniel.autoconfig.annotation.ConfigEntry;
 import java.util.ArrayList;
 import java.util.List;
 
-@Config(name = "fortress_tweak")
-public class FortressTweakConfig implements ConfigData {
+@Config(name = "nether_tweaks")
+public class NetherTweaksConfig implements ConfigData {
     @ConfigEntry.Category("Fortress")
     @ConfigEntry.BoundedDiscrete(min = -100, max = 100)
     public int fortressOffset = 48;