+++ src/main/java/net/funkpla/nether_tweaks/NetherTweaksMod.java
@@ -1,4 +1,4 @@
-package net.funkpla.fortress_tweak;
+package net.funkpla.nether_tweaks;
 
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
@@ -6,13 +6,13 @@ import net.fabricmc.api.ModInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class FortressTweakMod implements ModInitializer {
-    public static final String MOD_ID = "fortress_tweak";
+public class NetherTweaksMod implements ModInitializer {
+    public static final String MOD_ID = "nether_tweaks";
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 
     @Override
     public void onInitialize() {
         LOGGER.info("Nether Tweaks loaded.");
-        AutoConfig.register(FortressTweakConfig.class, JanksonConfigSerializer::new);
+        AutoConfig.register(NetherTweaksConfig.class, JanksonConfigSerializer::new);
     }
 }