+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-# Changelog for [Fortress Tweak](https://github.com/murphy-slaw/fortress-tweak)
+# Changelog for [Nether Tweaks](https://github.com/murphy-slaw/nether-tweaks)
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.1.0] 2025-06-03
+
+- renamed from Fortress Tweak to Nether Tweaks
+- Added option to restrict lighting nether portals to a configured list of structures
 
 ## [1.0.0] 2024-11-27